+++ message/FileRequestMessage.java
@@ -88,7 +88,7 @@ public ByteBuffer getRawMessage()
 	
 	// Write the message header
 	rawMessage.put(this.getMessageCode().getCode());
-	rawMessage.putLong((long)this.getRawMessageLength());
+	rawMessage.putInt(this.getRawMessageLength());
 	
 	// Write the filename
 	try
